fix(engine): LP64 long-width audit (F1–F6) + Array<T>::operator[] off-by-one
F1 pigtool.h: SYS_INT32/SYS_UINT32 use `int` not `long` under LP64
  (__LINUX__/__ANDROID__ guard); fixes 8-byte int32 on x86-64.

F2 oad.h: _oadHeader chunkId/chunkSize/version long → int32.

F3 mailbox.hp/cc: ReadMailbox/WriteMailbox params + _mailboxBase long → int32.

F4 movecam.cc: idxShot/shotIndex long → int32 (5 sites).

F5 color.hpi: temp local long → int in Color +/-/* operators.

F6 actor.cc/warp.cc: SPECIAL_COLLISION pointer stored/read through
  uintptr_t instead of long* (avoids truncation on LP64).

Array off-by-one (dormant since CVS rev 1.3, 2003-05-23):
  operator[](int32) used `if(index > _num)` — should be `>=`.
  Sequential init of an odd-N array left _num at N-1, so Size()
  under-reported by 1 and the last local mailbox slot missed the range
  check. SMB ?-block (NumberOfLocalMailboxes=13, odd) was first actor
  to exercise a non-zero odd local-mailbox array; mailbox 2012
  (SMB_QBLOCK_DIE) asserted on read. Fix: one char, array.hpi:195.

Also: SMB mailbox table reorganised (global → local slots for qblock
state), Generator actor fix (script now runs while idle), runtime-long
audit plan + investigation doc + BUGS.md entries, wf-edit OAD plumbing,
minor CMake/Lua/anim/memory/renderassets LP64 cleanup.

+++ engine/mutation/wfmut.cpp
@@ -408,7 +408,7 @@ bool SetMailbox(Level& level, ActorIdx idx, int mailboxIndex, double value)
         return fail("wfmut::SetMailbox: mailboxIndex must be >= 0");
     // Actor exposes its mailbox bank via GetMailboxes(); backing storage is
     // bounds-checked by MailboxesWithStorage in DBSTREAM builds.
-    actor->GetMailboxes().WriteMailbox(static_cast<long>(mailboxIndex), Scalar::FromDouble(value));
+    actor->GetMailboxes().WriteMailbox(mailboxIndex, Scalar::FromDouble(value));   // int -> int32 mailbox index
     ok();
     return true;
 }
@@ -419,7 +419,7 @@ std::optional<double> GetMailbox(const Level& level, ActorIdx idx, int mailboxIn
     if (!actor) return std::nullopt;
     if (mailboxIndex < 0)
         return failopt<double>("wfmut::GetMailbox: mailboxIndex must be >= 0");
-    Scalar v = actor->GetMailboxes().ReadMailbox(static_cast<long>(mailboxIndex));
+    Scalar v = actor->GetMailboxes().ReadMailbox(mailboxIndex);   // int -> int32 mailbox index
     ok();
     return static_cast<double>(v.AsFloat());
 }
